Move CV-only scoring off NCRFModel into _crossvalidation
eval_l2 and the compute_es_metric staticmethod were hosted on NCRFModel but
called from exactly one place, _score_mu, and encode cross-validation
scoring policy rather than general model behavior. Hosting them on the
frozen model leaked CV-specific API onto a class whose outside callers only
use eval_obj / explained_variance.

Move both to _crossvalidation.py as module functions taking the model (or
list of models), and call them directly from _score_mu. Behavior is
unchanged.
